+++ HISTORY.md
@@ -3,6 +3,8 @@
 ## not yet released, version 0.8.0
 
 - Implemented function `create` to create a new instance of typed-function.
+- Implemented a simple `typed.find` function to find the implementation of a
+  specific function signature.
 - Extended the error messages to denote the function name, like `"Too many 
   arguments in function foo (...)"`.